requirements.txt
streamlit
google-generativeai
vertexai
openai
pypdf
PyMuPDF
Pillow
